:atom: generalise classifiers
Generalise classifiers and make analysis run with key values taken from
a configuration file.

Also reduced the runtime by avoiding double computation of test values,
pulling the comparison out of the classifier functions and into the
classifier script.

To run, the classifier script currently needs to be copied to the base
directory and run as:

    python poas_ir_classifier.py -c configs/config.toml <image>